requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
django-environ==0.10.0
django-extensions==3.2.3
Django==4.2.1
djangorestframework==3.15.2
docopt==0.6.2
idna==3.4
Pillow==10.0.0
pipreqs==0.4.13
psycopg2-binary==2.9.6
requests==2.30.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
yarg==0.1.9
